Moving tests into src/tests/ so we don't have to deal with TypeScript project references.
