+++ src/bin/generate-test.ts
@@ -2,13 +2,13 @@ import 'source-map-support/register';
 import util from 'util';
 import DataDir from '../lib/data/dir';
 import { DataSet } from '../lib/data/set';
-import { abbrActionDetails, abbrEventDetails } from '../../tests/deal-generator/utils';
 import { Engine } from "../lib/engine/engine";
 import { Hubspot } from '../lib/hubspot';
 import { Logger } from '../lib/log';
 import { License } from '../lib/marketplace/model/license';
 import { Transaction } from '../lib/marketplace/model/transaction';
 import { engineConfigFromENV } from '../lib/parameters/env-config';
+import { abbrActionDetails, abbrEventDetails } from '../tests/deal-generator/utils';
 
 function TEMPLATE({ runDealGenerator, GROUP, RECORDS, EVENTS, ACTIONS }: any) {
   it(`describe test`, () => {
