+++ src/tests/deal-generator/basics.test.ts
