+++ src/tests/deal-generator/partner-transacted.test.ts
