+++ tsconfig.json
@@ -22,6 +22,5 @@
     "src/**/*.js",
     "src/**/*.ts",
     "src/**/*.d.ts",
-    "tests/**/*.ts",
   ]
 }
